Allign stage04 to new pipeline structure.
ATTENTION! There actually is a bug in merge_wave_definition rule.
It seems not to be possible to pass None (~) as an argumnet in the yaml file
to ADDITIONAL PROPERTIES parameter since it is used for iteration. Still, if
not set at None the merge_wave_definition rule requires an argument.

In order to temporary avoid the additional properies these are currently
not passed to the "merge_wave_definition" rule -> set at None by default.