Smart downsampling implemented in stage 02.
Non-homogeneous way to performa downsampling of input image according to the sigma to noise ratio.
Optimal macropixel dimension is thus explored in a tree search method according to the
distribution of signal intensities (a shapiro test is implemented, different analysis are planned
to be included in future work).

Three different exit condition are implemented:
- when at least N nodes in the same tree branch are proved to be normally distributed
- when N consecutive nodes are proved to be normally distributed
- when the fraction of normally distributed children is higher to a certain threshold

NB The output of this block is compliant to all blocks in stage 03 but it is not yet
usable to run stage04 and stage05 analysis. Allignment for at least some of the methods implemented
at these two further stages are planned to be implemented in the next steps.