[Fast Run] pipeline supporting also pkl format
Blocks outputs can be stored as both "nix" or "pkl" format.
Pikle happens to be much faster in the read/write process to a
consequent speed-up in the execution time.

To implement this  "load_input" and "write_output" functions have been
implemented in utils.io. These functions read in input the format chosen
by the user and process the input/output flow accordingly.